+++ src/assets/gulp/tasks/requireJs/unit-bundle.js
@@ -1,8 +1,0 @@
-/**
- * Gulp tasks for bundling RequireJS modules.
- */
-const gulp = require('gulp');
-
-gulp.task('unit-bundle', (cb) => {
-    cb();
-});