+++ src/assets/gulp/tasks/requireJs/util/modules.js
@@ -22,9 +22,7 @@ function createModuleConfig(filename, indexFile, uniteConfiguration, clientModul
     configContent.push('    }');
     configContent.push('});');
     
-    if (preloadModules.length > 0) {
-        configContent.push('var preloadModules = [' + preloadModules.join(',') + '];');
-    }
+    configContent.push('var preloadModules = [' + preloadModules.join(',') + '];');
 
     fs.writeFile(filename, configContent.join(os.EOL), (err) => {
         if (err) {