+++ src/configuration/models/unite/uniteModuleLoader.ts
@@ -2,6 +2,7 @@
  * Type for Unite Configuration moduleLoader.
  */
 export type UniteModuleLoader =
+    "Browserify" |
     "RequireJS" |
     "SystemJS" |
     "Webpack";